requirements.txt
openai
python-dotenv
gradient==3.1.0